backend/requirements.txt
django==3.2
python-dotenv==0.20
django-split-settings==1.1
gunicorn==20.0.4
psycopg2-binary==2.9
django_extensions==3.2.0
django-debug-toolbar==3.4.0